+++ src/typings/typedArrays.d.ts
@@ -1,0 +1,11 @@
+type TypedArray =
+  | Uint8Array
+  | Uint8ClampedArray
+  | Uint16Array
+  | Uint32Array
+  | BigUint64Array
+  | Int8Array
+  | Int32Array
+  | BigInt64Array
+  | Float32Array
+  | Float64Array;